requirements/prod.txt
Django==4.2.8
djangorestframework==3.14
python-decouple==3.8
drf-spectacular==0.26.1
psycopg2-binary==2.9.9
Pillow==10.1.0
gunicorn==21.2.0
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-filter==23.5
django-jazzmin==2.6.0
django-summernote==0.8.20.0